Implement a better way for admin checks
Admin checks require attributes from Admin object, and not user
object, so there was no point in retrieving their user object and
then having to fetch the admin object from the database.

Update deps.py, and create a `get_current_admin()` method which
directly returns the corresponding Admin object which allows us to
easily check the permissions of that admin.

Signed-off-by: Akhil Narang <me@akhilnarang.dev>

# Conflicts:
#	app/app/api/api_v1/endpoints/schools.py
#	app/app/api/deps.py
#	app/app/models/__init__.py