test utils: create_random_user(): allow setting admin permissions
While we do want a seemingly random user to be created, we do require
a very specific set of permissions, else the test *will* fail

Add permissions as an optional parameter, which is only checked when
`user_type` is set to admin. By default we create admins with a
permission of 0, so update that object to set the given permission.

This fixes our issue of creating an admin that has access to school
data specifically.

Signed-off-by: Akhil Narang <me@akhilnarang.dev>

# Conflicts:
#	app/app/tests/utils/user.py