schemas: year: move validations to YearBase
Change type of start_year and end_year in the validators
Optional[int] has been used since in the case of YearUpdate,
there are chances that these values will not be provided.

Update the checks to accordingly ensure that a value has been given,
and only then validate and update it.

Signed-off-by: Akhil Narang <me@akhilnarang.dev>