Don't use alembic for non MySQL setups
We've configured these revisions with and for MySQL, many of them
are not compatible with other RDBMS, like the PostgreSQL setup we
have to run in docker. Since that is basically just used for tests
in GitHub Actions, we always create a fresh DB anyway, so just
directly initialize it.

Also, remove the alembic upgrade step from prestart.sh, since again
that is just used by the above action.

Signed-off-by: Akhil Narang <me@akhilnarang.dev>