tests: utils: utils.py: Make a recursive `compare_api_and_db_query_results` function to assert equality of API call results and DB query results
The function essentially asserts if the value of a key in the `api_result` dict is the same as the value of that key in the `db_dict` dict. If the value happens to be a dict itself, and the value for that key in `db_dict` is a dict as well, then the function will recursively call itself to compare values in a similar fashion.

Signed-off-by: ksdfg <ksdfg123@gmail.com>