backend: modify our testing setup
Use our own base docker image and a smaller docker-compose setup

This brings our test runtime down from 5+ minutes to less than 1.5

This setup uses our base docker image, with the cloned repo as a
volume to run our lint checks, and also runs pytest in a compose
setup along with a postgres container.

Cleanup the workflow a bit, and adjust the environment variables
to account for the change in scripts.

Signed-off-by: Akhil Narang <me@akhilnarang.dev>