app/models: Switch to Association Proxy and Association Model from Assocciation Object for many-to-many relationship between Student and Division
This was needed because we had to store some additional fields in the common table for StudentDivision.

StudentDivision will be the common table between Student & Division, that will also store the batch_number for the student. This table will have relationships with both the models using the target attribute student_division
To connect this target collection to both Student & Divsion, we use association_proxy() property.
AssociationProxy is used to create a read/write "view" of a target attribute across a relationship.

The AssociationProxy object produced by the association_proxy() function is an instance of a Python descriptor. It is always declared with the user-defined class being mapped, regardless of whether Declarative or classical mappings via the mapper() function are used.
The proxy functions by operating upon the underlying mapped attribute or collection in response to operations, and changes made via the proxy are immediately apparent in the mapped attribute, as well as vice versa. The underlying attribute remains fully accessible.