tests: utils:to_json() - RIP logic. No clue how tf this fixes the problem
My guess is that it goes in recursion because since there is no backref for the original object, if a call for the original object appears in between other recursively called objects, the condition for stopping recursion fails and fall into infinite loop. (Very weird tbh)
Here is an example of what was happening with the Division object https://katb.in/paw1223.md

So added a check for the original table an addition to the backref check.